requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
yt-dlp
ffmpeg-python
python-dotenv
sqlalchemy
bcrypt
email-validator
